pyproject.toml
[tool.poetry]
name = "digital-library"
version = "0.1.0"
description = ""
authors = ["Yura Shurhay <yura.shurhay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.1"
django-cors-headers = "3.13.0"
django-extensions = "3.2.0"
django-filter = "22.1"
djangorestframework = "3.13.1"
djangorestframework-jwt = "1.11.0"
ipython = "8.3.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
gunicorn = "^22.0.0"
celery = "^5.4.0"
psycopg2 = "^2.9.9"
faker = "^26.0.0"
elasticsearch = "7.14.0"
elasticsearch-dsl = "7.4.0"
django-elasticsearch-dsl = "7.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
